`integer` means `int32` not `int64` (#378)
- As part of https://github.com/prisma/prisma-engines/pull/2970 we need to make a difference between `i32` and `i64` values in SQLite.
- In the past, we have `integer` as a "shortcut" for `int64` but this is not really the case, every db out there defines `int` as a 32 bit integer, not 64 bit.
- Because of this, a few tests and results need to be fixed:
  * OIDs return an unsigned 32 integer but this does not fit in `i32` so we return `i64` in this case.
  * Certain parameters, like `OFFSET` and `LIMIT`, are too unsigned integers, doing the same, explicitly using `i64`
- Adjusted some tests to reflect this

+++ src/ast/values.rs
@@ -225,9 +225,9 @@ impl<'a> Value<'a> {
     /// Creates a new 64-bit signed integer.
     pub fn integer<I>(value: I) -> Self
     where
-        I: Into<i64>,
+        I: Into<i32>,
     {
-        Value::Int64(Some(value.into()))
+        Value::Int32(Some(value.into()))
     }
 
     /// Creates a new decimal value.
